@@ -1,11 +1,15 @@
 package com.example.ecommercebe.config;
 
+import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Contact;
 import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.info.License;
+import io.swagger.v3.oas.models.security.SecurityRequirement;
+import io.swagger.v3.oas.models.security.SecurityScheme;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+
 
 @Configuration
 public class OpenAPIConfig {
@@ -21,11 +25,23 @@ public class OpenAPIConfig {
         Info info = new Info()
                 .title("Student Management API")
                 .version("1.0")
-                .description("This is a simple student management API")
-                .termsOfService("https://www.google.com")
                 .contact(contact)
+                .description("This API exposes endpoints to manage student.").termsOfService("https://google.com")
                 .license(mitLicense);
 
-        return new OpenAPI().info(info);
+        SecurityScheme securityScheme = new SecurityScheme()
+                .name("Authorization")
+                .type(SecurityScheme.Type.HTTP)
+                .scheme("bearer")
+                .bearerFormat("JWT")
+                .in(SecurityScheme.In.HEADER);
+
+        SecurityRequirement securityRequirement = new SecurityRequirement().addList("Authorization");
+
+        return new OpenAPI()
+                .info(info)
+                .addSecurityItem(securityRequirement)
+                .components(new Components().addSecuritySchemes("Authorization", securityScheme));
     }
+
 }
